Change notification feeds names in streams. Removed use of notification manager for the local stream service, which doesn't need it
