+++ cdap-common/src/main/java/co/cask/cdap/common/conf/Constants.java
@@ -593,9 +593,9 @@ public final class Constants {
      * Notifications in Streams constants.
      */
     public static final class Stream {
-      public static final String STREAM_HEARTBEAT_FEED_CATEGORY = "streamHeartbeats";
-      public static final String STREAM_HEARTBEAT_FEED_NAME = "streamHeartbeats";
-      public static final String STREAM_FEED_CATEGORY = "streams";
+      public static final String STREAM_FEED_CATEGORY = "stream";
+      public static final String STREAM_INTERNAL_FEED_CATEGORY = "stream.internal";
+      public static final String STREAM_HEARTBEAT_FEED_NAME = "heartbeat";
 
       /** Default number of bytes received by a stream after which a notification is sent */
       public static final long DEFAULT_DATA_THRESHOLD = 1024 * 1024 * 1024;
