+++ cdap-data-fabric/src/main/java/co/cask/cdap/data/stream/service/AbstractStreamService.java
@@ -16,11 +16,7 @@
 
 package co.cask.cdap.data.stream.service;
 
-import co.cask.cdap.common.conf.Constants;
 import co.cask.cdap.data.stream.StreamCoordinatorClient;
-import co.cask.cdap.notifications.feeds.NotificationFeed;
-import co.cask.cdap.notifications.feeds.NotificationFeedException;
-import co.cask.cdap.notifications.feeds.NotificationFeedManager;
 import com.google.common.util.concurrent.AbstractIdleService;
 import com.google.common.util.concurrent.Service;
 
@@ -31,7 +27,6 @@ public abstract class AbstractStreamService extends AbstractIdleService implemen
 
   private final StreamCoordinatorClient streamCoordinatorClient;
   private final StreamFileJanitorService janitorService;
-  private final NotificationFeedManager feedManager;
 
   /**
    * Children classes should implement this method to add logic to the start of this {@link Service}.
@@ -48,16 +43,13 @@ public abstract class AbstractStreamService extends AbstractIdleService implemen
   protected abstract void doShutdown() throws Exception;
 
   protected AbstractStreamService(StreamCoordinatorClient streamCoordinatorClient,
-                                  StreamFileJanitorService janitorService,
-                                  NotificationFeedManager feedManager) {
+                                  StreamFileJanitorService janitorService) {
     this.streamCoordinatorClient = streamCoordinatorClient;
     this.janitorService = janitorService;
-    this.feedManager = feedManager;
   }
 
   @Override
   protected final void startUp() throws Exception {
-    createHeartbeatsFeed();
     streamCoordinatorClient.startAndWait();
     janitorService.startAndWait();
     initialize();
@@ -68,24 +60,5 @@ public abstract class AbstractStreamService extends AbstractIdleService implemen
     doShutdown();
     janitorService.stopAndWait();
     streamCoordinatorClient.stopAndWait();
-  }
-
-  /**
-   * Create Notification feed for stream's heartbeats, if it does not already exist.
-   */
-  private void createHeartbeatsFeed() throws NotificationFeedException {
-    // TODO worry about namespaces here. Should we create one heartbeat feed per namespace?
-    NotificationFeed streamHeartbeatsFeed = new NotificationFeed.Builder()
-      .setNamespace(Constants.DEFAULT_NAMESPACE)
-      .setCategory(Constants.Notification.Stream.STREAM_HEARTBEAT_FEED_CATEGORY)
-      .setName(Constants.Notification.Stream.STREAM_HEARTBEAT_FEED_NAME)
-      .setDescription("Streams heartbeats feed.")
-      .build();
-
-    try {
-      feedManager.getFeed(streamHeartbeatsFeed);
-    } catch (NotificationFeedException e) {
-      feedManager.createFeed(streamHeartbeatsFeed);
-    }
   }
 }
