+++ cdap-data-fabric/src/main/java/co/cask/cdap/data/stream/service/DistributedStreamService.java
@@ -27,6 +27,8 @@ import co.cask.cdap.common.zookeeper.coordination.ResourceModifier;
 import co.cask.cdap.common.zookeeper.coordination.ResourceRequirement;
 import co.cask.cdap.data.stream.StreamCoordinatorClient;
 import co.cask.cdap.data.stream.StreamLeaderListener;
+import co.cask.cdap.notifications.feeds.NotificationFeed;
+import co.cask.cdap.notifications.feeds.NotificationFeedException;
 import co.cask.cdap.notifications.feeds.NotificationFeedManager;
 import com.google.common.base.Function;
 import com.google.common.base.Supplier;
@@ -63,6 +65,7 @@ public class DistributedStreamService extends AbstractStreamService {
   private final DiscoveryServiceClient discoveryServiceClient;
   private final StreamMetaStore streamMetaStore;
   private final ResourceCoordinatorClient resourceCoordinatorClient;
+  private final NotificationFeedManager feedManager;
   private final Set<StreamLeaderListener> leaderListeners;
   private Supplier<Discoverable> discoverableSupplier;
 
@@ -77,18 +80,20 @@ public class DistributedStreamService extends AbstractStreamService {
                                   DiscoveryServiceClient discoveryServiceClient,
                                   StreamMetaStore streamMetaStore,
                                   Supplier<Discoverable> discoverableSupplier,
-                                  NotificationFeedManager notificationFeedManager) {
-    super(streamCoordinatorClient, janitorService, notificationFeedManager);
+                                  NotificationFeedManager feedManager) {
+    super(streamCoordinatorClient, janitorService);
     this.zkClient = zkClient;
     this.discoveryServiceClient = discoveryServiceClient;
     this.streamMetaStore = streamMetaStore;
     this.discoverableSupplier = discoverableSupplier;
+    this.feedManager = feedManager;
     this.resourceCoordinatorClient = new ResourceCoordinatorClient(zkClient);
     this.leaderListeners = Sets.newHashSet();
   }
 
   @Override
   protected void initialize() throws Exception {
+    createHeartbeatsFeed();
     resourceCoordinatorClient.startAndWait();
     coordinationSubscription = resourceCoordinatorClient.subscribe(discoverableSupplier.get().getName(),
                                                                    new StreamsLeaderHandler());
@@ -140,6 +145,25 @@ public class DistributedStreamService extends AbstractStreamService {
         }
       }
     };
+  }
+
+  /**
+   * Create Notification feed for stream's heartbeats, if it does not already exist.
+   */
+  private void createHeartbeatsFeed() throws NotificationFeedException {
+    // TODO worry about namespaces here. Should we create one heartbeat feed per namespace?
+    NotificationFeed streamHeartbeatsFeed = new NotificationFeed.Builder()
+      .setNamespace(Constants.DEFAULT_NAMESPACE)
+      .setCategory(Constants.Notification.Stream.STREAM_INTERNAL_FEED_CATEGORY)
+      .setName(Constants.Notification.Stream.STREAM_HEARTBEAT_FEED_NAME)
+      .setDescription("Stream heartbeats feed.")
+      .build();
+
+    try {
+      feedManager.getFeed(streamHeartbeatsFeed);
+    } catch (NotificationFeedException e) {
+      feedManager.createFeed(streamHeartbeatsFeed);
+    }
   }
 
   /**
