+++ cdap-data-fabric/src/main/java/co/cask/cdap/data/stream/service/LocalStreamService.java
@@ -17,7 +17,6 @@
 package co.cask.cdap.data.stream.service;
 
 import co.cask.cdap.data.stream.StreamCoordinatorClient;
-import co.cask.cdap.notifications.feeds.NotificationFeedManager;
 import com.google.inject.Inject;
 
 /**
@@ -27,9 +26,8 @@ public class LocalStreamService extends AbstractStreamService {
 
   @Inject
   public LocalStreamService(StreamCoordinatorClient streamCoordinatorClient,
-                            StreamFileJanitorService janitorService,
-                            NotificationFeedManager notificationFeedManager) {
-    super(streamCoordinatorClient, janitorService, notificationFeedManager);
+                            StreamFileJanitorService janitorService) {
+    super(streamCoordinatorClient, janitorService);
   }
 
   @Override
